Chapter 9 Update

+++ Chapter 9/ProductService/Sources/App/Controllers/ProductController.swift
@@ -12,19 +12,13 @@ final class ProductsController {
         }
         if let query = try request.query.get(String?.self, at: "query") {
             querybuilder.filter(\.$name ~~ query)
-            let query1: String = query
-            querybuilder.filter(.group(.or) {
+            querybuilder.group(.or) {
                 $0.filter(\Product.$name ~~ query).filter(\Product.$description ~~ query)
-            })
-            
-            querybuilder.filter(.group([
-            .value(.field(path: ["name"], schema: Product.schema, alias: nil), .contains(inverse: false, .anywhere), .bind(query)),
-            .value(.field(path: ["description"], schema: Product.schema, alias: nil), .contains(inverse: false, .anywhere), .bind(query))
-            ], .or))
+            }
         }
         if let idsString = try request.query.get(String?.self, at: "ids") {
-            let ids:[Int?] = idsString.split(separator: ",").map { Int(String($0)) }
-            querybuilder.filter(\.$id.field ~~ ids)
+            let ids:[Int] = idsString.split(separator: ",").map { Int(String($0)) ?? 0 }
+            querybuilder.filter(\.$id ~~ ids)
         }
         
         return querybuilder.all().map { products in
@@ -42,30 +36,38 @@ final class ProductsController {
     
     func edit(_ request: Request)throws -> EventLoopFuture<ProductResponse> {
         let input = try request.content.decode(ProductInput.self)
-        let id = try request.query.get(Int?.self)
-        
-        return Product.query(on: request.db).filter(\.$id == id).all().flatMap { products in
-            if products.count == 0 {
-                return request.eventLoop.makeFailedFuture(Abort(.badRequest, reason: "No product. found!"))
+        if let id = request.parameters.get("id", as: Int.self)
+        {
+            return Product.query(on: request.db).filter(\.$id == id).all().flatMap { products in
+                if products.count == 0 {
+                    return request.eventLoop.makeFailedFuture(Abort(.badRequest, reason: "No product. found!"))
+                }
+                let product = products.first!
+                
+                product.name = input.name
+                product.description = input.description
+                product.price = input.price
+                product.categoryId = input.categoryId
+                
+                return product.save(on: request.db).map { _ in
+                    return ProductResponse(id: product.id!, name: product.name, description: product.description, price: product.price)
+                }
             }
-            let product = products.first!
-            
-            product.name = input.name
-            product.description = input.description
-            product.price = input.price
-            product.categoryId = input.categoryId
-            
-            return product.save(on: request.db).map { _ in
-                return ProductResponse(id: product.id!, name: product.name, description: product.description, price: product.price)
-            }
+        }
+        else {
+            throw Abort(.badRequest, reason: "No id given.")
         }
     }
     
     func delete(_ request: Request)throws -> EventLoopFuture<HTTPStatus> {
-        let id = try request.query.get(Int?.self)
-        
-        return Product.query(on: request.db).filter(\.$id == id).delete().map { _ in
-            return .ok
+        if let id = request.parameters.get("id", as: Int.self)
+        {
+            return Product.query(on: request.db).filter(\.$id == id).delete().map { _ in
+                return .ok
+            }
+        }
+        else {
+            throw Abort(.badRequest, reason: "No id given.")
         }
     }
